Better Guesstimation
Seems if there's no valid versions, we don't even bother picking a terrible version. So, now if there's only one, we'll assume that one is the one we'd want to edit if so pressed.

This fixes a "bad URL" when you attempt to edit the size of a "No Versions" entry.

Also, seems I broke the newlines by editing this file.